requirements.txt
streamlit>=1.15.0
requests>=2.28.0
python-docx>=0.8.11
openpyxl>=3.0.10
python-pptx>=0.6.21
pdfminer.six>=20221105
pandas>=1.3.0